Add happo integration
Through trial-and-error, this is the minimal commit I could come up with
in order to get an end-to-end happo run to work.

I'm not using the happo-plugin-typescript plugin here even though it's
meant to make typescript integration simpler. The default react-scripts
webpack config has all the necessary setup, so we can just use it
directly.

The NODE_ENV=development flag is needed for the right webpack config to
be generated by react-scripts. Without that flag, the function
invocation (craWebpackConfig()) fails with an error message about
missing NODE_ENV.